refactor: move links to constants
