+++ src/lib/constants.ts
@@ -1,13 +1,3 @@
-export const routes = {
-  BLOG: '/blog',
-  PROJECTS: '/projects',
-  SNIPPETS: '/snippets',
-  ENDORSEMENTS: '/endorsements',
-  ABOUT: '/about',
-  STATS: '/stats',
-  RESUME: '/resume',
-}
-
 export const MAX_VIEWS_PER_SESSION = 30
 export const MAX_SHARES_PER_SESSION = 10
 export const MAX_REACTIONS_PER_SESSION = 20
